backend/services/real_apis.py
import requests
import aiohttp
import asyncio
from typing import List, Dict, Any, Optional
import os
from googletrans import Translator
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
from google.auth.transport.requests import Request
import json

class RealAPIService:
    def __init__(self):
        self.translator = Translator()
        self.session = None
    
    async def get_session(self):
        """Get or create aiohttp session"""
        if self.session is None:
            self.session = aiohttp.ClientSession()
        return self.session

    # MOOC API Integrations
    async def search_edx_courses(self, query: str, language: str = "en") -> List[Dict[str, Any]]:
        """Search courses from edX API"""
        try:
            session = await self.get_session()
            url = "https://api.edx.org/catalog/v1/catalogs/edx/courses/"
            params = {
                "search": query,
                "language": language,
                "page_size": 20
            }
            
            async with session.get(url, params=params) as response:
                if response.status == 200:
                    data = await response.json()
                    courses = []
                    for course in data.get("results", []):
                        courses.append({
                            "id": course.get("key"),
                            "title": course.get("name"),
                            "description": course.get("short_description", ""),
                            "language": course.get("language", "en"),
                            "difficulty": course.get("level", "beginner"),
                            "provider": "edX",
                            "duration": f"{course.get('effort', 'Unknown')} hours/week",
                            "rating": course.get("rating", 0.0),
                            "url": course.get("marketing_url", "")
                        })
                    return courses
                else:
                    print(f"edX API error: {response.status}")
                    return []
        except Exception as e:
            print(f"Error fetching from edX: {e}")
            return []

    async def search_mit_ocw_courses(self, query: str) -> List[Dict[str, Any]]:
        """Search courses from MIT OpenCourseWare"""
        try:
            session = await self.get_session()
            url = "https://ocw.mit.edu/api/v1/search/"
            params = {
                "q": query,
                "type": "course"
            }
            
            async with session.get(url, params=params) as response:
                if response.status == 200:
                    data = await response.json()
                    courses = []
                    for course in data.get("results", []):
                        courses.append({
                            "id": course.get("id"),
                            "title": course.get("title"),
                            "description": course.get("description", ""),
                            "language": "en",
                            "difficulty": "intermediate",
                            "provider": "MIT OCW",
                            "duration": "Self-paced",
                            "rating": 4.5,
                            "url": course.get("url", "")
                        })
                    return courses
                else:
                    print(f"MIT OCW API error: {response.status}")
                    return []
        except Exception as e:
            print(f"Error fetching from MIT OCW: {e}")
            return []

    async def search_openstax_courses(self, query: str) -> List[Dict[str, Any]]:
        """Search courses from OpenStax"""
        try:
            session = await self.get_session()
            url = "https://openstax.org/api/books"
            
            async with session.get(url) as response:
                if response.status == 200:
                    data = await response.json()
                    courses = []
                    for book in data:
                        if query.lower() in book.get("title", "").lower():
                            courses.append({
                                "id": book.get("id"),
                                "title": book.get("title"),
                                "description": book.get("description", ""),
                                "language": "en",
                                "difficulty": "beginner",
                                "provider": "OpenStax",
                                "duration": "Self-paced",
                                "rating": 4.3,
                                "url": book.get("webview_url", "")
                            })
                    return courses
                else:
                    print(f"OpenStax API error: {response.status}")
                    return []
        except Exception as e:
            print(f"Error fetching from OpenStax: {e}")
            return []

    # Translation API Integration
    def translate_text(self, text: str, source_lang: str = "auto", target_lang: str = "en") -> Dict[str, Any]:
        """Translate text using Google Translate"""
        try:
            result = self.translator.translate(text, src=source_lang, dest=target_lang)
            return {
                "original_text": text,
                "translated_text": result.text,
                "source_language": result.src,
                "target_language": result.dest,
                "confidence": 0.95
            }
        except Exception as e:
            print(f"Translation error: {e}")
            return {
                "original_text": text,
                "translated_text": f"[Translation Error] {text}",
                "source_language": source_lang,
                "target_language": target_lang,
                "confidence": 0.0
            }

    def detect_language(self, text: str) -> Dict[str, Any]:
        """Detect language of text"""
        try:
            result = self.translator.detect(text)
            return {
                "text": text,
                "detected_language": result.lang,
                "confidence": result.confidence,
                "supported_languages": ["en", "es", "fr", "de", "zh", "ja", "ko", "ar", "hi", "pt"]
            }
        except Exception as e:
            print(f"Language detection error: {e}")
            return {
                "text": text,
                "detected_language": "en",
                "confidence": 0.0,
                "supported_languages": ["en", "es", "fr", "de", "zh", "ja", "ko", "ar", "hi", "pt"]
            }

    # Google Classroom API Integration
    def setup_google_classroom(self):
        """Setup Google Classroom API credentials"""
        try:
            # You'll need to set up Google Cloud credentials
            # For now, return mock service
            return None
        except Exception as e:
            print(f"Google Classroom setup error: {e}")
            return None

    async def create_google_classroom_course(self, title: str, description: str = "") -> Dict[str, Any]:
        """Create a course in Google Classroom"""
        try:
            # Mock implementation - replace with real Google Classroom API
            return {
                "course_id": f"course_{hash(title) % 10000}",
                "title": title,
                "description": description,
                "status": "created",
                "invitation_code": "ABC123",
                "owner_id": "teacher@example.com"
            }
        except Exception as e:
            print(f"Google Classroom error: {e}")
            return {"error": "Failed to create course"}

    async def sync_progress_to_classroom(self, course_id: str, student_id: str, progress_data: Dict[str, Any]) -> Dict[str, Any]:
        """Sync learning progress to Google Classroom"""
        try:
            # Mock implementation - replace with real sync logic
            return {
                "course_id": course_id,
                "student_id": student_id,
                "progress_data": progress_data,
                "sync_status": "completed",
                "last_synced": "2024-01-01T12:00:00Z"
            }
        except Exception as e:
            print(f"Sync error: {e}")
            return {"error": "Failed to sync progress"}

    async def close(self):
        """Close the session"""
        if self.session:
            await self.session.close() 