pyproject.toml
[project]
name = "Multilingual-Learning-Path-Generator"
version = "0.1.0"
description = "An agentic application that generates adaptive, multilingual study plans from MOOC catalogs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core MCP and LangChain
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "mcp-use>=1.3.3",
    "mcp[cli]>=1.10.1",
    
    # Backend Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    
    # Frontend
    "streamlit>=1.35.0",
    "plotly>=5.18.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.15.0",
    "sqlite3",
    
    # HTTP and API
    "httpx>=0.28.0",
    "aiohttp>=3.10.0",
    "requests>=2.32.0",
    
    # Translation and NLP
    "googletrans>=4.0.0",
    "langdetect>=1.0.9",
    
    # Google APIs
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.120.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    "loguru>=0.7.0",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true